Coordinate should be double not int. Also, adapt getCoordinate for future purpose. Calculate area formula will need next and previous coordinates through iterations

+++ src/main/java/Coordinate.java
@@ -1,19 +1,19 @@
-/**
+/*
  * Created by asivolella on 2016-01-27.
  */
 public class Coordinate {
-    private int x, y;
+    private double x, y;
 
-    public Coordinate(int x, int y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
